+++ src/contractual_conditions/dto/create-contractual_conditions.dto.ts
@@ -11,7 +11,7 @@ import {
   EstadoContrato,
   Periodicidad,
   TipoContrato,
-} from '../entities/contractualConditions.entity';
+} from '../entities/contractual_conditions.entity';
 import { Transform } from 'class-transformer';
 
 export class CreateContractualConditionDto {
