+++ src/contractual_conditions/entities/contractual_conditions.entity.ts
